Refactor implementations of InStream to have outcome remapping.
The outcome remapping is used in a new InStream method called "quit",
which has the same semantics as Outcome.quit, but applies the remapping
to the requested outcome type.

As a result, there is now a single FileInStream which is used for
all three stream types, and which accepts an outcome remapping in a form
of Map<Outcome.Type, Outcome.Type>, where empty mappings are interpreted
as "leave it as is". There is no need for three implementing classes, so
they are removed.

As there are two remappings currently used: the empty one and the one
that replaces every non-OK outcome with FAIL, the non-trivial remapping
is implemented as a static final field in Outcome.